+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4]
+
+- Configuration is now done with a spreadsheet instead of yaml
+- JSON examples for the LLM are now generated automatically from the spreadsheet
+- added some extra methods to make the Classifier object easier to work with
+- allows overriding the cache when you want to test prompts
+- better testing for prompt parsing and generation
+
 ## [0.1.3]
 
 - Flattened collation